// frontend/src/generated/models/liquidityWorkspaceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { LiquidityWorkspaceResponseAssurance } from './liquidityWorkspaceResponseAssurance';
import type { LiquidityWorkspaceResponseCashForecast } from './liquidityWorkspaceResponseCashForecast';
import type { WorkspaceContextResponse } from './workspaceContextResponse';
import type { LiquidityWorkspaceResponseCovenantsItem } from './liquidityWorkspaceResponseCovenantsItem';
import type { LiquidityWorkspaceResponseDebtItem } from './liquidityWorkspaceResponseDebtItem';
import type { LiquidityWorkspaceResponseLineage } from './liquidityWorkspaceResponseLineage';
import type { LiquidityWorkspaceResponseMetricsItem } from './liquidityWorkspaceResponseMetricsItem';
import type { LiquidityWorkspaceResponseStressesItem } from './liquidityWorkspaceResponseStressesItem';
import type { LiquidityWorkspaceResponseWorkingCapitalItem } from './liquidityWorkspaceResponseWorkingCapitalItem';

export interface LiquidityWorkspaceResponse {
  as_of: string;
  assurance?: LiquidityWorkspaceResponseAssurance;
  cash_forecast?: LiquidityWorkspaceResponseCashForecast;
  context: WorkspaceContextResponse;
  covenants?: LiquidityWorkspaceResponseCovenantsItem[];
  debt?: LiquidityWorkspaceResponseDebtItem[];
  lineage?: LiquidityWorkspaceResponseLineage;
  metrics?: LiquidityWorkspaceResponseMetricsItem[];
  projection_version: number;
  source_snapshot_ids?: string[];
  stresses?: LiquidityWorkspaceResponseStressesItem[];
  working_capital?: LiquidityWorkspaceResponseWorkingCapitalItem[];
}