// frontend/src/generated/models/profitabilityWorkspaceResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * CFO Command Center API
 * OpenAPI spec version: 0.3.0
 */
import type { ProfitabilityWorkspaceResponseAllocationAssurance } from './profitabilityWorkspaceResponseAllocationAssurance';
import type { ProfitabilityWorkspaceResponseAssurance } from './profitabilityWorkspaceResponseAssurance';
import type { WorkspaceContextResponse } from './workspaceContextResponse';
import type { ProfitabilityWorkspaceResponseLineage } from './profitabilityWorkspaceResponseLineage';
import type { ProfitabilityWorkspaceResponseMarginWaterfallItem } from './profitabilityWorkspaceResponseMarginWaterfallItem';
import type { ProfitabilityWorkspaceResponseMetricsItem } from './profitabilityWorkspaceResponseMetricsItem';
import type { ProfitabilityWorkspaceResponseProfitabilityMatrixItem } from './profitabilityWorkspaceResponseProfitabilityMatrixItem';
import type { ProfitabilityWorkspaceResponseSegmentsItem } from './profitabilityWorkspaceResponseSegmentsItem';
import type { ProfitabilityWorkspaceResponseSensitivitySummaryItem } from './profitabilityWorkspaceResponseSensitivitySummaryItem';

export interface ProfitabilityWorkspaceResponse {
  allocation_assurance?: ProfitabilityWorkspaceResponseAllocationAssurance;
  as_of: string;
  assurance?: ProfitabilityWorkspaceResponseAssurance;
  context: WorkspaceContextResponse;
  lineage?: ProfitabilityWorkspaceResponseLineage;
  margin_waterfall?: ProfitabilityWorkspaceResponseMarginWaterfallItem[];
  metrics?: ProfitabilityWorkspaceResponseMetricsItem[];
  profitability_matrix?: ProfitabilityWorkspaceResponseProfitabilityMatrixItem[];
  projection_version: number;
  segments?: ProfitabilityWorkspaceResponseSegmentsItem[];
  sensitivity_summary?: ProfitabilityWorkspaceResponseSensitivitySummaryItem[];
  source_snapshot_ids?: string[];
}